2005/01/02 02:37:18+01:00 handhelds.org!CoreDump
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded
into handhelds.org:/home/mhentges/OpenEmbedded/bitbake/openembedded

2005/01/02 02:37:06+01:00 handhelds.org!CoreDump
rootfs_ipk: Check the logfile for error messages and error-out if necessary

2005/01/02 02:34:54+01:00 handhelds.org!CoreDump
opie-image: Fail gracefully if ipkg.conf is missing

BKrev: 41d750f5kYn8z0gnIKOxG2q7gzJaXw

+++ packages/meta/opie-image.bb
@@ -31,7 +31,13 @@ merge_feeds() {
 
 	if ! test -z "${FEED_URIS}"
 	then
-
+		# Die gracefully if ipkg-collateral failed
+		if ! test -e "${IMAGE_ROOTFS}/etc/ipkg.conf"
+		then
+			echo "[${IMAGE_ROOTFS}/etc/ipkg.conf] is missing!"
+			exit 1
+		fi
+		
 		# comment out existing feed-sources inserted by ipkg-collateral
 		cat ${IMAGE_ROOTFS}/etc/ipkg.conf | sed "s/^src\ /#src\ /" > ${IMAGE_ROOTFS}/etc/ipkg.conf_
 		rm ${IMAGE_ROOTFS}/etc/ipkg.conf && mv ${IMAGE_ROOTFS}/etc/ipkg.conf_ ${IMAGE_ROOTFS}/etc/ipkg.conf
